Waid-backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
aiofiles
requests
python-docx
reportlab
PyPDF2
pymupdf
pdfminer.six
